+++ src/graphql_clj/validator.clj
@@ -12,6 +12,7 @@
             [graphql-clj.validator.rules.unique-variable-names]
             [graphql-clj.validator.rules.unique-operation-names]
             [graphql-clj.validator.rules.unique-input-field-names]
+            [graphql-clj.validator.rules.unique-fragment-names]
             [graphql-clj.visitor :as visitor]
             [graphql-clj.spec :as spec]
             [instaparse.core :as insta]
@@ -32,7 +33,8 @@
             graphql-clj.validator.rules.no-fragment-cycles/rules
             graphql-clj.validator.rules.fragments-on-composite-types/rules
             graphql-clj.validator.rules.unique-variable-names/rules
-            graphql-clj.validator.rules.unique-operation-names/rules]))
+            graphql-clj.validator.rules.unique-operation-names/rules
+            graphql-clj.validator.rules.unique-fragment-names/rules]))
 
 (defn- validate [visit-fn]
   (try (visit-fn)